+++ packages/app/src/components/presentational/CreateFyghterButton/CreateFyghterButton.spec.tsx
@@ -1,13 +1,13 @@
 import * as React from "react";
 import { shallow } from "enzyme";
 import * as FyghtState from "../../../state";
-import { FyghterCreationModal } from ".";
+import { CreateFyghterButton } from ".";
 
-describe("FyghterCreationModal", () => {
+describe("CreateFyghterButton", () => {
   test("should render the component", () => {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     jest.spyOn(FyghtState, "useFyghtState").mockImplementation((): any => [{}, { createFyghter: jest.fn() }]);
 
-    expect(shallow(<FyghterCreationModal />)).toMatchSnapshot();
+    expect(shallow(<CreateFyghterButton />)).toMatchSnapshot();
   });
 });
