--- packages/app/src/components/presentational/FyghterCreationModal/index.tsx
@@ -1,1 +1,0 @@
-export * from "./FyghterCreationModal";
